Replace Prototype with JQuery, RJS with JSON, use UJS
This replaces the ugly RJS code that will no longer be supported in
Rails 3, and replaces it with a simple JSON code that uses something
similar to an sexp.

In addition, the usage of Prototype is replaced with JQuery and
JQuery-Autocomplete, leading to simpler code.

Finally, instead of hard coding the JS handlers in the HTML tag, use
unobtrusive javascript to add them.