Get spam ready for deployment on Heroku
Add Gemfile and Procfile as required to run on Heroku Cedar.
Because we don't actually want to use bundler to run the application,
comment out the bundler lines from boot.rb and application.rb.

To get the logging to work, automatically set it to log to stdout
if an environment variable is present.

While here, upgrade Rails version to 3.2.8.