Precompile all views and freeze template caches
This uses the new template precompilation features in Roda 3.4.0.
Make sure to use a consistent string or symbol when loading
templates.

With these changes, the application does not need access to the
public directory if the public directory is served by a front-end
webserver.

Include requested pages in the ajax test webserver log. Allow
running the ajax test with a pledged application.