Switch to using transactional testing in all cases
For the ajax tests with a real browser, use Sequel 5.81's
new multithreaded transactional testing support.  In order
for this to work, switch from unicorn to puma as the test
webserver, and run the webserver in the same process, but
in a different thread.

Start using permissions_policy Roda plugin.

Add logger to Gemfile, to avoid warnings/errors in Ruby 3.4/3.5.